remove Twitch4JHandler, move object translation into from method on ChatMessage
